requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
pandas==1.4.4
plotly==5.9.0
Requests==2.30.0
scikit_learn==1.0.2
streamlit==1.21.0
